refactor: Stubbing용 객체 패키지 분리
