+++ src/test/java/nextstep/subway/unit/LoginServiceTest.java
@@ -7,12 +7,11 @@ import nextstep.member.application.LoginService;
 import nextstep.member.application.dto.GithubAccessTokenRequest;
 import nextstep.member.application.dto.GithubAccessTokenResponse;
 import nextstep.member.application.dto.TokenRequest;
-import nextstep.member.domain.GithubResponses;
+import nextstep.member.domain.stub.GithubResponses;
 import nextstep.member.domain.MemberRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
